+++ Assets/UniGLTF/Editor/UniGLTF/ExportDialog/MeshExportValidator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 15b6e58b6aafb284783d551ac475ca3f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
